tests/wpt/web-platform-tests/webxr/webxr-supported-by-feature-policy.html

<!DOCTYPE html>
<title>Test that xr is advertised in the feature list</title>
<!-- The Feature Policy feature name(s) for WebXR are still TBD. See the issue below. This test checks what is currently in the spec. -->
<link rel="help" href="https://github.com/immersive-web/webxr/issues/308">
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://immersive-web.github.io/webxr/#feature-policy">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('xr', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise xr.');
</script>
